feat: add SearchButton component for contextual search functionality
- Introduced a new SearchButton component that opens a contextual search modal based on the current route (SRC20 or Stamp).
- Implemented keyboard shortcut (Ctrl/Cmd + S) for quick access to the search functionality.
- Enhanced user experience with tooltip visibility management on mouse events.
- Refactored existing SearchSRC20Modal and SearchStampModal to support the new search mechanism.
